+++ syft/core/frameworks/torch/tensor/__init__.py
@@ -1,0 +1,14 @@
+from syft.core.frameworks.torch.tensor.syft_tensor import _SyftTensor
+from syft.core.frameworks.torch.tensor.torch_object import _TorchObject
+from syft.core.frameworks.torch.tensor.fixed_precision_tensor import _FixedPrecisionTensor
+from syft.core.frameworks.torch.tensor.generalized_pointer_tensor import _GeneralizedPointerTensor
+from syft.core.frameworks.torch.tensor.local_tensor import _LocalTensor
+from syft.core.frameworks.torch.tensor.mpc_tensor import _MPCTensor
+from syft.core.frameworks.torch.tensor.plus_is_minus_tensor import _PlusIsMinusTensor
+from syft.core.frameworks.torch.tensor.pointer_tensor import _PointerTensor
+from syft.core.frameworks.torch.tensor.torch_tensor import _TorchTensor
+from syft.core.frameworks.torch.tensor.torch_variable import _TorchVariable
+
+__all__ = ['_SyftTensor', '_LocalTensor',
+           '_PointerTensor', '_FixedPrecisionTensor', '_TorchTensor',
+           '_PlusIsMinusTensor', '_GeneralizedPointerTensor', '_MPCTensor']
