+++ syft/core/frameworks/torch/tensor/fixed_precision_tensor.py
@@ -1,0 +1,6 @@
+from syft.core.frameworks.torch.tensor import _SyftTensor
+
+class _FixedPrecisionTensor(_SyftTensor):
+
+    def __init__(self, child, parent, torch_type, owner=None):
+        super().__init__(child=child, parent=parent, torch_type=torch_type, owner=owner)
