+++ syft/core/frameworks/torch/tensor/plus_is_minus_tensor.py
@@ -1,0 +1,49 @@
+import torch
+
+import syft as sy
+from syft.core.frameworks.torch.tensor import _SyftTensor
+
+class _PlusIsMinusTensor(_SyftTensor):
+    """
+    Example of a custom overloaded _SyftTensor
+
+    Role:
+    Converts all add operations into sub/minus ones.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    # The table of command you want to replace
+    substitution_table = {
+        'torch.add': 'torch.add'
+    }
+
+    class overload_functions:
+        """
+        Put here the functions you want to overload
+        Beware of recursion errors.
+        """
+        @staticmethod
+        def add(x, y):
+            return x.add(y)
+
+        @staticmethod
+        def get(attr):
+            attr = attr.split('.')[-1]
+            return getattr(sy._PlusIsMinusTensor.overload_functions, attr)
+
+    # Put here all the methods you want to overload
+    def add(self, arg):
+        """
+        Overload the add method and execute another function or method with the provided args
+        """
+        _response = self.sub(arg)
+
+        return _response
+
+    def abs(self):
+        """
+        Overload the abs() method and execute another function
+        """
+        return torch.abs(self)
